Create theme file and other small tweaks
- Separate out the theme into its own file
- Fix typo
- Move a button to a more logical place in the code (the css places
  it correctly regardless so the code placement is just helpful to
  those reading the code)
